services/loader/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-dataplex==2.*
google-cloud-storage==2.*
google-cloud-pubsub==2.*
cloudevents==1.*
